refactor: 단체 지정 도메인 분리
